+++ notebooks/104_data_preprocess_scenarios_ar6.py
@@ -109,7 +109,7 @@ else:
         "emissions": "primap-202503",
         "gdp": "wdi-2025",
         "population": "un-owid-2025",
-        "gini": "unu-wider-2025",
+        "gini": "wdi-2025",
         "target": "pathway",
     }